requirements.txt
altgraph==0.17.2
asgiref==3.5.2
crispy-forms-materialize==0.2
dj-database-url==1.0.0
Django==4.1
django-baton==2.4.1
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-taggit==3.0.0
django-taggit-forms==0.1.1
djangorestframework==3.14.0
future==0.18.2
gunicorn==20.1.0
pefile==2022.5.30
Pillow==9.2.0
psycopg2==2.9.5
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.8
python-decouple==3.6
pytz==2022.6
pywin32-ctypes==0.2.0
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0
